% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{annotate_snps_binary}
\alias{annotate_snps_binary}
\title{Assigns SNPs with annotations based on overlap}
\usage{
annotate_snps_binary(sumstats, annotations, keep.annot.only = TRUE)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics}

\item{annotations}{Paths to annotation BED files}

\item{keep.annot.only}{Logical. If TRUE, only returns the
snp and annotation columns}
}
\value{
if \code{keep.annot.only=TRUE}, returns
a data frame or tibble of snp and annotations. Otherwise, returns
a data frame or tibble of GWAS summary statistics and annotations.
}
\description{
Assigns SNPs with annotations based on overlap
}
