% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotator}
\alias{annotator}
\title{Each annotation gets assigned SNPs based on overlap}
\usage{
annotator(gwas, annotations)
}
\arguments{
\item{annotations}{}
}
\description{
Each annotation gets assigned SNPs based on overlap
}
