% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotator_merged}
\alias{annotator_merged}
\title{Annotations for causal SNPs (apply these after fine-mapping!)}
\usage{
annotator_merged(sumstats, annotations)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics}

\item{annotations}{Paths to annotation BED files}
}
\description{
Annotations for causal SNPs (apply these after fine-mapping!)
}
