% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{assign_snp_locus}
\alias{assign_snp_locus}
\title{Assigns each SNP to one LD block}
\usage{
assign_snp_locus(cleaned.sumstats, LD_Blocks)
}
\arguments{
\item{cleaned.sumstats}{A data frame of GWAS summary statistics.}

\item{LD_Blocks}{A data frame of LD blocks}
}
\description{
Assigns each SNP to one LD block
}
