% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_functions.R
\name{calc_stress}
\alias{calc_stress}
\title{Calculate stress between plates}
\usage{
calc_stress(world, spread = 5)
}
\arguments{
\item{world}{A world list; the output of \code{\link{gen_plates}}}

\item{spread}{How many tiles away from the plate border should stress
start being felt? Defaults to 5.}
}
\value{
an updated world list, where the map contains a stress column
}
\description{
Calculate stress between plates
}
