% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_functions.R
\name{classify_heights}
\alias{classify_heights}
\title{break stress by quantiles}
\usage{
classify_heights(world, topography)
}
\arguments{
\item{world}{a world object, e.g. the output of \code{\link{gen_world}}.}

\item{topography}{a topography table containing at least an id column and a
q column (with the minimum quantile for each terrain id).}
}
\value{
an updated world object
}
\description{
break stress by quantiles
}
