% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_functions.R
\name{classify_temps}
\alias{classify_temps}
\title{break temperatures by quantiles}
\usage{
classify_temps(world, breaks)
}
\arguments{
\item{world}{a world object, e.g. the output of \code{\link{gen_world}}.}

\item{breaks}{a named vector of quantile breaks to apply. The names are used
as labels. E.g. c("A" = 0, "B" = 0.5) would divide the colder half of the
world as "A", and the warmer half of the world as "B".}
}
\value{
an updated world object
}
\description{
break temperatures by quantiles
}
