% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compile_structure_plot_data}
\alias{compile_structure_plot_data}
\title{Compile input data for making structure plot}
\usage{
compile_structure_plot_data(mat, categories)
}
\arguments{
\item{mat}{matrix of input data}

\item{categories}{annotation categories}
}
\value{
a data frame for making structure plot
}
\description{
this function is adapted from the 'fastTopics' package
https://stephenslab.github.io/fastTopics/
}
