% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{compute_distance_weight}
\alias{compute_distance_weight}
\title{Find genes around SNPs and assign weights based on distance}
\usage{
compute_distance_weight(
  snp.ranges,
  gene.ranges,
  d0 = 50000,
  maxgap = 1e+06,
  type = c("promoter", "gene")
)
}
\arguments{
\item{snp.ranges}{A GRanges object with the SNP locations.}

\item{gene.ranges}{A GRanges object with the gene or promoter locations.}

\item{d0}{A scaling parameter used for computing weight based on SNP-gene distance.
Weight = exp(-dist/d0). Default = 50000 (50kb).}

\item{maxgap}{Max distance to find genes around SNPs. Default = 1e6 (1Mb)}

\item{type}{Are the gene.ranges promoters or gene locations?}
}
\value{
A GRanges object with SNP to gene distance and weights calculated based on distance.
}
\description{
Find genes around SNPs and assign weights based on distance
}
