% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{compute_gene_pip_by_hierarchies}
\alias{compute_gene_pip_by_hierarchies}
\title{Map SNPs to genes and assign weights}
\usage{
compute_gene_pip_by_hierarchies(
  finemap.gr,
  genomic.annots,
  intron.mode = TRUE,
  active_promoter_method = c("OCRs", "pcHiC", "ABC", "distance", "none"),
  enhancer_loop_method = "ABC.pcHiC.nearby20kb",
  c.dist = 1e+05,
  dist.to = c("end", "tss", "midpoint"),
  cols.to.keep = c("snp", "chr", "pos", "ref", "alt", "locus", "pip", "gene_name",
    "category", "weight")
)
}
\arguments{
\item{finemap.gr}{Genomic Ranges of finemapping result.}

\item{genomic.annots}{A list of genomic annotations.}

\item{intron.mode}{If FALSE, do not assign intronic SNPs to genes containing the introns.}

\item{active_promoter_method}{Method to define active promoters.}

\item{enhancer_loop_method}{Enhancer loop method}

\item{c.dist}{A scaling number used for computing weight based on SNP-gene distance. Weight = exp(-dist/c). Default = 1e5 (100kb).}

\item{dist.to}{Compute distance to TSS, gene midpoint, or end (right point).}

\item{cols.to.keep}{columns to keep in the SNP gene weights}
}
\description{
Map SNPs to genes and assign weights
}
