% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{extract_gene_level_result}
\alias{extract_gene_level_result}
\title{Extract gene level result from SNP level gene mapping result}
\usage{
extract_gene_level_result(snp.gene.pip.res, gene.annots)
}
\arguments{
\item{snp.gene.pip.res}{A data frame of SNP level gene mapping result}

\item{gene.annots}{a GRange object of gene annotations}
}
\description{
Extract gene level result from SNP level gene mapping result
}
