% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{filter_coaccess}
\alias{filter_coaccess}
\title{Filtering coaccessibility data}
\usage{
filter_coaccess(coaccess, cor.thresh = 0.7, dist.thresh = 1e+06)
}
\arguments{
\item{coaccess}{data frame of coaccessibility}

\item{cor.thresh}{Threshold for coaccessibility}

\item{dist.thresh}{Threshold for distance}
}
\description{
Filtering coaccessibility data
}
