% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{find_nearest_genes}
\alias{find_nearest_genes}
\title{Find the nearest genes for top SNPs in each locus.}
\usage{
find_nearest_genes(
  gwas.gr,
  genes.gr,
  dist.to = c("genebody", "tss"),
  cols.to.keep = c("snp", "chr", "pos", "nearest_gene")
)
}
\arguments{
\item{gwas.gr}{a GRanges object of the GWAS summary statistics}

\item{genes.gr}{a GRanges object of gene information}

\item{dist.to}{Find nearest genes by distance to gene body or TSS.}

\item{cols.to.keep}{columns to keep in the returned GRanges object}
}
\value{
a data frame with SNP location and nearest gene.
}
\description{
Find the nearest genes for top SNPs in each locus.
}
