% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{finemapping_annot_trackplot}
\alias{finemapping_annot_trackplot}
\title{Make gene track plot using Gviz}
\usage{
finemapping_annot_trackplot(
  finemapstats,
  region,
  gene.annots,
  bigSNP,
  txdb,
  genome = c("hg19", "hg38"),
  genetrack.db = c("txdb", "gene.annots", "UCSC"),
  filter.protein.coding.genes = TRUE,
  countsdata,
  peaks,
  loops,
  filter.loop.genes = NULL,
  filter.loop.snps = NULL,
  data.colors = seq_along(countsdata),
  data.ylim = c(0, 1),
  color.pip.by = c("locus", "cs", "none"),
  highlight.snps = NULL,
  highlight.colors = "pink",
  genelabel.side = c("above", "right", "left", "below"),
  track.sizes = NULL,
  rotation.title = 90,
  verbose = FALSE
)
}
\arguments{
\item{finemapstats}{A GRanges object or data frame of finemapping summary statistics}

\item{region}{A GRanges object or data frame for the genomic range to plot}

\item{gene.annots}{A GRanges object of gene annotations}

\item{bigSNP}{A bigsnpr object attached via bigsnpr::snp_attach()}

\item{txdb}{A txdb object of gene annotations}

\item{genome}{Genome assembly version, hg19 (default) or hg38.}

\item{genetrack.db}{Select a gene annotation database to use. Options:
\code{txdb}: use the \code{txdb} objec.
\code{gene.annots}: use the \code{gene.annots} object.
\code{UCSC} uses \verb{UCSC knownGene} annotations.}

\item{filter.protein.coding.genes}{If TRUE, only shows protein coding gene}

\item{countsdata}{A list of counts data}

\item{peaks}{A list of peaks}

\item{loops}{A list of chromatin loops, e.g. PC-HiC, ABC, etc.}

\item{filter.loop.genes}{If TRUE, only shows HiC loops connected to
the gene(s)}

\item{filter.loop.snps}{If TRUE, only shows HiC loops connected to
the SNP(s)}

\item{data.colors}{Colors for the \code{countsdata} tracks}

\item{data.ylim}{ylim range for the \code{countsdata} tracks}

\item{color.pip.by}{color SNPs in the PIP track by \code{locus}, \code{cs},
or \code{none} (same color).}

\item{highlight.snps}{SNPs (rsIDs) to highlight}

\item{highlight.colors}{Colors for the highlighted SNPs}

\item{genelabel.side}{Side to put gene labels,
options are: above (default), right, left, below}

\item{track.sizes}{Sizes of the tracks}

\item{rotation.title}{Rotation of the track titles}

\item{verbose}{if TRUE, print detail messages for plotting}
}
\description{
Make gene track plot using Gviz
}
