% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{gen_world}
\alias{gen_world}
\title{wrapper to generate a world}
\usage{
gen_world(
  n_plates,
  spread,
  weight = 1,
  noise = 50,
  map_x,
  map_y = map_x,
  gravity_range = c(0.2, 1),
  height_range = c(-10, 10),
  dist_method = c("square", "manhattan")
)
}
\arguments{
\item{n_plates}{The number of plates in the map. Can be either 1 value
for a world generated from a single map, or a vector of values for a world
generated of several tectonic maps superimposed.}

\item{spread}{How many tiles away from the plate border should stress
start being felt? Must be a vector of the same length as n_plates}

\item{weight}{when more than one map is generated, how strongly should
subsequent maps imprint on the first? defaults to 1; i.e. sub-maps fully
imprint on first map.}

\item{noise}{Strength of the noise to add to the heights at the end of the
process. noise is the standard deviation of the normal distribution from
which the values are drawn. Set to 0 to not add noise.}

\item{map_x}{map x dimension in tiles}

\item{map_y}{map y dimension in tiles}
}
\value{
A world object.
}
\description{
Uses one or more sets of plates to create a complex surface.
}
