% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{gene_by_distance}
\alias{gene_by_distance}
\title{Find all genes (promoters) within 1MB genes and assign weights based on distance}
\usage{
gene_by_distance(
  snps.gr,
  promoters.gr,
  c.dist = 1e+05,
  dist.to = c("tss", "midpoint", "end")
)
}
\arguments{
\item{snps.gr}{Genomic Ranges with the SNP locations.}

\item{promoters.gr}{Genomic Ranges with the gene promoter locations.}

\item{c.dist}{A scaling number used for computing weight based on SNP-gene distance.
Weight = exp(-dist/c). Default = 1e5 (100kb).}
}
\description{
Find all genes (promoters) within 1MB genes and assign weights based on distance
}
