% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gene_manhattan_plot}
\alias{gene_manhattan_plot}
\title{Make gene manhattan plot}
\usage{
gene_manhattan_plot(
  gene.pip,
  chr = "chr",
  pos = "pos",
  gene = "gene_name",
  pip = "gene_pip",
  highlight = FALSE,
  annotate = FALSE,
  sig.pip = 0.5,
  annotate.pip = NULL,
  highlight.genes = NULL,
  ylim = NULL,
  title = "",
  max.overlaps = 20
)
}
\arguments{
\item{gene.pip}{Gene level finemapping result}

\item{chr}{Name of the chr column in the gene finemapping summary statistics data}

\item{gene}{Name of the gene name column in the gene finemapping summary statistics data}

\item{pip}{Name of the gene PIP column in the gene finemapping summary statistics data}

\item{sig.pip}{Signficance cutoff of gene PIP}

\item{annotate.pip}{annotate genes with gene PIP > annotate.pip}

\item{highlight.genes}{Genes to highlight}

\item{ylim}{Truncate gene PIP to ylim value in the plot}

\item{title}{Title of the plot}

\item{max.overlaps}{Limit of overlapping labels (ggrepel setting)}

\item{POS}{Name of the pos column in the gene finemapping summary statistics data}
}
\description{
Make gene manhattan plot
}
