% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{get_locus_level_gene_pip}
\alias{get_locus_level_gene_pip}
\title{Get locus level gene PIP}
\usage{
get_locus_level_gene_pip(gene.mapping.res)
}
\arguments{
\item{gene.mapping.res}{A data frame of SNP-level gene mapping result}
}
\description{
Get locus level gene PIP
}
