% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{get_nearby_interactions}
\alias{get_nearby_interactions}
\title{Get nearby interactions for enhancer regions near promoters}
\usage{
get_nearby_interactions(enhancers, promoters, max.dist = 20000)
}
\arguments{
\item{enhancers}{A GRanges object of enhancer regions}

\item{promoters}{A GRanges object of promoters}

\item{max.dist}{Max distance betweeen enhancer regions and promoters (default: 20kb)}
}
\value{
A GRanges object of nearby interactions
}
\description{
Get nearby interactions for enhancer regions near promoters
}
