% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_UKBB_region_info}
\alias{get_UKBB_region_info}
\title{Get region info with filenames of LD matrices and variant information}
\usage{
get_UKBB_region_info(
  LD_Blocks,
  LDREF.dir,
  prefix = "ukb_b37_0.1",
  LD_matrix_ext = "RDS",
  snp_info_ext = "Rvar"
)
}
\arguments{
\item{LD_Blocks}{A data frame of LD blocks}

\item{LDREF.dir}{Directory of UKBB LD reference files}

\item{prefix}{prefix name of the UKBB LD reference files}

\item{LD_matrix_ext}{File extension of LD matrix files}

\item{snp_info_ext}{File extension of SNP information files}
}
\value{
A data frame with information of the variants in the LD matrix.
}
\description{
Get region info with filenames of LD matrices and variant information
}
