% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{LD_diagnosis_susie_rss}
\alias{LD_diagnosis_susie_rss}
\title{Perform diagnosis to check the consistency between the z-scores and LD matrix,
and detect problematic z-scores (e.g. allele switch issue) using
diagnostic functions from SuSiE RSS.}
\usage{
LD_diagnosis_susie_rss(z, R, n)
}
\arguments{
\item{z}{z scores.}

\item{R}{LD correlation matrix.}

\item{n}{The sample size. (Optional, but highly recommended.)}
}
\value{
a data frame of expected z-scores and test statistics
from \code{susieR::kriging_rss}.
}
\description{
Perform diagnosis to check the consistency between the z-scores and LD matrix,
and detect problematic z-scores (e.g. allele switch issue) using
diagnostic functions from SuSiE RSS.
}
