% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{LDblock_view_summary}
\alias{LDblock_view_summary}
\title{LD block view summary table}
\usage{
LDblock_view_summary(genemapping_res, finemap.gr)
}
\arguments{
\item{genemapping_res}{data frame of gene mapping result}

\item{finemap.gr}{GRange object of fine mapping result}
}
\description{
LD block view summary table
}
