% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_ABC}
\alias{load_ABC}
\title{Load ABC scores and save as a GRanges object}
\usage{
load_ABC(ABC.file, ABC.thresh = 0, full.element = TRUE, flank = 0)
}
\arguments{
\item{ABC.file}{ABC file downloaded from Nasser et al.
(https://www.engreitzlab.org/resources).}

\item{ABC.thresh}{Significance threshold of ABC scores (default = 0).}

\item{full.element}{Logitic; if TRUE, use full length of ABC elements.
Otherwise, use the narrow regions provided in the ABC data.}

\item{flank}{Flanking regions around ABC elements (default = 0).}
}
\description{
Load ABC scores and save as a GRanges object
}
