% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_finemapping_sumstat}
\alias{load_finemapping_sumstat}
\title{Load finemapping summary statistics data}
\usage{
load_finemapping_sumstat(
  finemapping.file,
  SNP = "SNP",
  CHR = "CHR",
  POS = "BP",
  PVAL = "pval",
  CS = "cs"
)
}
\arguments{
\item{finemapping.file}{Filename of finemapping summary statistics data}

\item{SNP}{Name of the SNP ID (rsID) column in the finemapping summary statistics data}

\item{CHR}{Name of the CHR column in the finemapping summary statistics data}

\item{POS}{Name of the POS column in the finemapping summary statistics data}

\item{PVAL}{Name of the P-value column in the finemapping data frame}

\item{CS}{Name of the CS column in the finemapping data frame}
}
\description{
Load finemapping summary statistics data
}
