% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_genomic_annots}
\alias{load_genomic_annots}
\title{Load genomic annotations}
\usage{
load_genomic_annots(genomic.annots.file)
}
\arguments{
\item{genomic_annots_file}{Genomic annotation rds file}
}
\description{
Load genomic annotations
}
