% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_narrowpeaks}
\alias{load_narrowpeaks}
\title{Load ENCODE narrow peak data and convert to a GRanges object}
\usage{
load_narrowpeaks(peak.file)
}
\arguments{
\item{peak.file}{ENCODE narrow peak file}
}
\description{
Load ENCODE narrow peak data and convert to a GRanges object
}
