% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pcHiC}
\alias{load_pcHiC}
\title{Load pcHiC data and save as a GRanges object}
\usage{
load_pcHiC(pcHiC.file)
}
\arguments{
\item{pcHiC.file}{pcHiC file downloaded from Jung et al.
(https://pubmed.ncbi.nlm.nih.gov/31501517/)}
}
\description{
Load pcHiC data and save as a GRanges object
}
