% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{load_UKBB_LDREF}
\alias{load_UKBB_LDREF}
\title{Load UKBB LD reference matrix and variant information}
\usage{
load_UKBB_LDREF(LD_Blocks, locus, LDREF.dir, prefix = "ukb_b37_0.1")
}
\arguments{
\item{LD_Blocks}{A data frame of LD blocks}

\item{locus}{locus ID}

\item{LDREF.dir}{Directory of UKBB LD reference files}

\item{prefix}{prefix name of the UKBB LD reference files}
}
\value{
A list, containing LD (correlation) matrix R and
a data frame with information of the variants in the LD matrix.
}
\description{
Load UKBB LD reference matrix and variant information
}
