% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.R
\name{make_genomic_annots}
\alias{make_genomic_annots}
\title{Make genomic annotations from a GTF file}
\usage{
make_genomic_annots(gtf_file, promoter.dist = 2000L)
}
\arguments{
\item{gtf_file}{Path to the GTF file}

\item{promoter.dist}{Bases (default: 2000) upstream of transcription start site as promoter regions}
}
\value{
A list of GRanges objects of genomic annotations
}
\description{
Make genomic annotations (genes, exons, introns, UTRs, promoters)
from a GTF file.
}
