% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_structure_plot}
\alias{make_structure_plot}
\title{Make a structure plot}
\usage{
make_structure_plot(
  mat,
  categories,
  colors,
  ticks = NULL,
  font.size = 9,
  highlight = NULL
)
}
\arguments{
\item{mat}{matrix of input data}

\item{categories}{annotation categories}

\item{colors}{Colors of the structure plot categories}

\item{ticks}{Labels of x-axis}

\item{font.size}{Font size of structure plot}

\item{highlight}{Highlight a sample}
}
\description{
Makeing a structure plot.
This function is adapted from the 'fastTopics' package
https://stephenslab.github.io/fastTopics/
}
