% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgen-package.R
\docType{package}
\name{mapgen}
\alias{mapgen-package}
\alias{mapgen}
\title{mapgen}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hugomflavio/mapgen}
  \item Report bugs at \url{https://github.com/hugomflavio/mapgen/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
