% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{match_gwas_bigsnp}
\alias{match_gwas_bigsnp}
\title{Match GWAS with bigSNP reference panel.}
\usage{
match_gwas_bigsnp(sumstats, bigSNP, strand_flip = TRUE, match.min.prop = 0.1)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics}

\item{bigSNP}{bigSNP object}

\item{strand_flip}{Whether to try to flip strand? (default is TRUE).
If so, ambiguous alleles A/T and C/G are removed.}

\item{match.min.prop}{Minimum proportion of variants in the smallest data
to be matched, otherwise stops with an error. Default: 10\%}
}
\description{
Match GWAS with bigSNP reference panel.
}
