% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_gwas_LDREF}
\alias{match_gwas_LDREF}
\title{Match GWAS sumstats with LD reference files. Only keep variants included in
LD reference.}
\usage{
match_gwas_LDREF(sumstats, R, snp_info)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics.}

\item{R}{LD matrix}

\item{snp_info}{Variant information for the LD matrix.}
}
\value{
A list, containing matched GWAS summary statistics and LD matrix.
}
\description{
Match GWAS sumstats with LD reference files. Only keep variants included in
LD reference.
}
