% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.bigsnp.gwas}
\alias{merge.bigsnp.gwas}
\title{merge.bigsnp.gwas}
\usage{
\method{merge}{bigsnp.gwas}(gwas, bigSNP)
}
\arguments{
\item{gwas}{GWAS summary statistics}

\item{bigSNP}{bigSNP object}
}
\description{
merge.bigsnp.gwas
}
