% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearby_interactions}
\alias{nearby_interactions}
\title{Get nearby interactions for enhancer regions near promoters}
\usage{
nearby_interactions(enhancer_regions, promoters, max.dist = 20000)
}
\arguments{
\item{enhancer_regions}{A GRanges object of enhancer regions}

\item{promoters}{A GRanges object of promoters}

\item{max.dist}{Max distance between enhancer regions and promoters (default: 20kb)}
}
\value{
A GRanges object of nearby interactions
}
\description{
Get nearby interactions for enhancer regions near promoters
}
