% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{partition_pip_annots}
\alias{partition_pip_annots}
\title{Partition PIPs into functional annotation categories.}
\usage{
partition_pip_annots(finemapstats, annots.list)
}
\arguments{
\item{finemapstats}{a GRanges object of fine mapping summary statistics}

\item{annots.list}{a list of GRanges objects of ordered functional annotation categories,
if a SNP is in multiple annotation categories,
it will be assigned to the first ordered category.}
}
\value{
a list with a data frame with the sum of PIPs and
a data frame with the number of SNPs included for each annotation category.
}
\description{
Partition PIPs into functional annotation categories.
}
