% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{partition_pip_regions}
\alias{partition_pip_regions}
\title{Partition PIPs into functional annotation regions}
\usage{
partition_pip_regions(finemap.gr, annots, unassigned = "Intergenic")
}
\arguments{
\item{finemap.gr}{fine mapping result}

\item{annots}{functional annotation regions}

\item{unassigned}{name of unassigned category, default: 'Intergenic'.}
}
\description{
Partition PIPs into functional annotation regions
}
