% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{prepare_torus_input_files}
\alias{prepare_torus_input_files}
\title{Prepare TORUS input files}
\usage{
prepare_torus_input_files(
  sumstats,
  annotation_bed_files,
  torus_annot_file = "torus_annotations.txt.gz",
  torus_zscore_file = "torus_zscore.txt.gz"
)
}
\arguments{
\item{sumstats}{cleaned summary statistics}

\item{annotation_bed_files}{annotation files in BED format.
The bed file must have three columns: chr, start, end.
Chromosomes should be numeric (no "chr")
and should be in hg19/b37 format.
You will get wrong results if you use hg38 or other
(The reference panel we used is in hg19).}

\item{torus_annot_file}{Annotation file name.}

\item{torus_zscore_file}{z-score file name.}
}
\value{
a list containing paths to the z-score file and annotation file.
}
\description{
Prepares two files necessary for running TORUS:
z-score file and annotation file.
}
\examples{
\dontrun{
torus.files <- prepare_torus_input_files(sumstats, annotation_bed_files)
}
}
