% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{PrepareTorusFiles}
\alias{PrepareTorusFiles}
\title{PrepareTorusFiles}
\usage{
PrepareTorusFiles(cleaned_sumstats, bed_annotations)
}
\arguments{
\item{cleaned_sumstats}{cleaned summary statistics from RunCleaner or other}

\item{bed_annotation}{a directory containing annotations in bed format. The bed file must have three columns: chr, start, end. Chromosomes should be numeric (no "chr")
and should be in hg19/b37 format. You will get wrong results if you use hg38 or other (reference panel is hg19).}
}
\value{
location to two files: torus zscore file and annotation file
}
\description{
Prepares two files necessary for running torus: z-score file and annotations file
}
