% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{process_finemapping_sumstat}
\alias{process_finemapping_sumstat}
\title{Process fine-mapping summary statistics data}
\usage{
process_finemapping_sumstat(
  finemap,
  pip.thresh = 1e-05,
  filterCS = FALSE,
  maxCS = 10
)
}
\arguments{
\item{finemap}{A data frame of fine-mapping summary statistics}

\item{pip.thresh}{PIP threshold (default = 1e-5).}

\item{filterCS}{If TRUE, limiting to SNPs within credible sets.}

\item{maxCS}{Maximum number of credible sets (default = 10).}
}
\description{
Process fine-mapping summary statistics data
}
