% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{process_gwas_sumstats}
\alias{process_gwas_sumstats}
\title{Process GWAS summary statistics and harmonize with LD reference}
\usage{
process_gwas_sumstats(
  sumstats,
  chr = "chr",
  pos = "pos",
  beta = "beta",
  se = "se",
  a0 = "a0",
  a1 = "a1",
  snp = "snp",
  pval = "pval",
  remove_indels = TRUE,
  LD_Blocks = NULL,
  bigSNP = NULL,
  LD_snp_info = NULL,
  strand_flip = TRUE,
  remove_strand_ambig = TRUE,
  ...
)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics.}

\item{chr}{Name of the chromosome column in summary statistics.}

\item{pos}{Name of the position column (base pair position).}

\item{beta}{Name of beta column (if you have Odds Ratio,
you will need to transform it to log(Odds Ratio)).}

\item{se}{Name of the standard error (se) column.}

\item{a0}{Column name of the reference allele.}

\item{a1}{Column name of the association/effect allele.}

\item{snp}{Name of the SNP ID (rsID) column.}

\item{pval}{Name of the p-value column.}

\item{remove_indels}{If TRUE, remove indels}

\item{LD_Blocks}{A data frame of LD blocks with four columns,
'chr', 'start', 'end', and 'locus' (locus ID).}

\item{bigSNP}{a \code{bigsnpr} object attached via \code{bigsnpr::snp_attach()}
containing the reference genotype panel.}

\item{LD_snp_info}{a data frame, SNP info in the LD reference,
with columns 'chrom', 'id', 'pos', 'alt', 'ref', 'locus'.}

\item{strand_flip}{Whether to flip signs when reverse complement matches? (default is TRUE).}

\item{remove_strand_ambig}{Whether to remove ambiguous alleles (A/T and C/G)? (default is TRUE).}
}
\value{
A data frame of processed GWAS summary statistics.
}
\description{
Process GWAS summary statistics and adds metadata, including:
the index in the bigSNP object of each SNP, and the LD block locus.
}
