% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_loop_data}
\alias{process_loop_data}
\title{Process chromatin loop data and save as a GRanges object}
\usage{
process_loop_data(loops, gene.annots = NULL, score.thresh = 0, flank = 0)
}
\arguments{
\item{loops}{A data frame of chromatin loops, with columns:
"chr", "start", "end", "gene_name", and "score" (optional).}

\item{gene.annots}{If provided, restrict to genes in gene.annots}

\item{score.thresh}{Numeric. Threshold of interaction scores.
(default = 0).}

\item{flank}{Integer. Extend bases on both sides of the regulatory elements
(default = 0).}
}
\value{
A GRanges object with processed chromatin loops,
with genomic coordinates of the regulatory elements and gene names.
}
\description{
Process chromatin loop data and save as a GRanges object
}
