% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_narrowpeaks}
\alias{process_narrowpeaks}
\title{Load ENCODE narrow peak data and convert to a GRanges object}
\usage{
process_narrowpeaks(peak.file)
}
\arguments{
\item{peak.file}{ENCODE narrow peak file}
}
\value{
a GRanges object for the peaks
}
\description{
Load ENCODE narrow peak data and convert to a GRanges object
}
