% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_LD_SNP_info}
\alias{read_LD_SNP_info}
\title{Read all SNP info in LD reference}
\usage{
read_LD_SNP_info(region_info)
}
\arguments{
\item{region_info}{A data frame of region information,
paths of LD matrices (R, correlation matrices),
and paths of SNP information files corresponding to the LD matrices.}
}
\value{
A data frame of SNP information in the LD reference
}
\description{
Read all SNP info in LD reference
}
