% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_finemapping}
\alias{run_finemapping}
\title{Run fine-mapping}
\usage{
run_finemapping(
  sumstats,
  R,
  bigSNP,
  priortype = c("torus", "uniform"),
  L = 1,
  estimate_residual_variance = FALSE
)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics; must have header!}

\item{bigSNP}{a bigsnpr object attached via bigsnpr::snp_attach()}

\item{priortype}{prior type: "torus" or "uniform".}

\item{L}{Number of causal signals}

\item{estimate_residual_variance}{The default is FALSE,
the residual variance is fixed to 1 or variance of y.
If the in-sample LD matrix is provided,
we recommend setting estimate_residual_variance = TRUE.}
}
\value{
list of finemapping results; one per LD block
}
\description{
Run finemapping with SuSiE for all LD blocks
}
