% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{run_gwas_cleaner}
\alias{run_gwas_cleaner}
\title{Clean GWAS summary statistics and adds metadata}
\usage{
run_gwas_cleaner(sumstats_file, cols.to.keep, bigSNP, LD_Blocks)
}
\arguments{
\item{sumstats_file}{file containing raw summary statistics.
Coordinates should be hg39/b37; the pipeline does not support hg38.}

\item{cols.to.keep}{character vector of the following columns:
chr, position, allele1, allele2, beta, se, unique id, pvalue}

\item{bigSNP}{bigSNP object.}

\item{LD_Blocks}{Reference LD blocks}
}
\value{
Cleaned summary statistics + LD block of every SNP,
as well as its index in the reference panel of genotypes
}
\description{
Cleans GWAS summary statistics and adds metadata
}
