% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_susie_rss}
\alias{run_susie_rss}
\title{Run fine-mapping with SuSiE using summary statistics (susie_rss)}
\usage{
run_susie_rss(
  sumstats,
  bigSNP,
  R,
  n,
  L = 1,
  useprior = FALSE,
  estimate_residual_variance = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sumstats}{summary statistics}

\item{bigSNP}{bigSNP object}

\item{R}{p x p correlation (LD) matrix}

\item{n}{The sample size}

\item{L}{Number of causal signals}

\item{useprior}{Logical, if TRUE, use the \code{torus_prior} column
in \code{sumstats} as prior.}

\item{estimate_residual_variance}{The default is FALSE,
the residual variance is fixed to 1 or variance of y.
If the in-sample LD matrix is provided,
we recommend setting estimate_residual_variance = TRUE.}

\item{verbose}{If verbose = TRUE, print progress,
and a summary of the optimization settings from susie.}
}
\value{
susie_rss output
}
\description{
Run fine-mapping with SuSiE using summary statistics (susie_rss)
}
