% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{RunCleaner}
\alias{RunCleaner}
\title{RunCleaner}
\usage{
RunCleaner(sumstats, ColsToKeep, bigSNP)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics. Coordinates should be hg39/b37; the pipeline does not support hg38.}

\item{ColsToKeep}{character vector of the following columns: chr, position, allele1, allele2, beta, se, unique id, pvalue}
}
\value{
Cleaned summary statistics + LD block of every SNP, as well as its index in the reference panel of genotypes
}
\description{
Cleans GWAS summary statistics and adds metadata
}
