% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{RunFinemapping}
\alias{RunFinemapping}
\title{RunFinemapping}
\usage{
RunFinemapping(sumstats, bigSNP, priortype = "torus")
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics; must have header!}

\item{bigSNP}{a bigsnpr object attached via bigsnpr::snp_attach()}

\item{priortype}{string; one of "torus" or "uniform"}
}
\value{
list of finemapping results; one per LD block
}
\description{
Runs SuSiE with L = 1
}
