% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{RunTorus}
\alias{RunTorus}
\title{RunTorus}
\usage{
RunTorus(
  torus_annot_file,
  torus_zscore_file,
  TORUS = system.file("torus", package = "finemappeR")
)
}
\value{
list of enrichment results and PIP of each SNP (both tibbles)
}
\description{
executes Torus
}
