% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{RunTorusFDR}
\alias{RunTorusFDR}
\title{RunTorusFDR}
\usage{
RunTorusFDR(
  torus_annot_file,
  torus_zscore_file,
  TORUS = system.file("torus", package = "finemappeR")
)
}
\value{
tibble containing fdr of each LD block/chunk
}
\description{
Runs Torus in FDR mode to estimate the FDR of each chunk containing a causal variant
}
