% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_functions.R
\name{scale_stress}
\alias{scale_stress}
\title{Scale stress values}
\usage{
scale_stress(world)
}
\arguments{
\item{world}{A world list containing stress values. The output of
\code{\link{calc_stress}}.}
}
\value{
an updated world list where the stress data has been scaled.
}
\description{
\code{\link{calc_stress}} can generate different ranges of values for
different parameters. This function scales stress to roughly -1000 to 1000.
}
