% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{snp_view_summary}
\alias{snp_view_summary}
\title{SNP view summary table}
\usage{
snp_view_summary(
  genemapping_res,
  gene.annots,
  finemapstats.gr,
  fractional.PIP.thresh = 0.02
)
}
\arguments{
\item{genemapping_res}{data frame of gene mapping result}

\item{gene.annots}{data frame of gene annotations}

\item{finemapstats.gr}{GRange object of fine mapping result}

\item{fractional.PIP.thresh}{Filter SNPs with fractional PIP cutoff (default: 0.02)}
}
\description{
SNP view summary table
}
