% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\title{structure plot}
\usage{
structure_plot(dat, colors, ticks = NULL, font.size = 9, highlight = NULL)
}
\arguments{
\item{dat}{a data frame obtained from @rdname compile_structure_plot_data}

\item{colors}{Colors of the structure plot categories}

\item{ticks}{Labels of x-axis}

\item{font.size}{Font size of structure plot}

\item{highlight}{Highlight a sample}
}
\description{
structure plot
}
