% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_annots.R
\name{substract_exons}
\alias{substract_exons}
\title{Substract all exon coordinates from all gene coordinates to get all intron coordinates}
\usage{
substract_exons(genes.gr, exons.gr)
}
\arguments{
\item{genes.gr}{GRange of gene coorindates}

\item{exons.gr}{GRange of exon coorindates}
}
\description{
Substract all exon coordinates from all gene coordinates to get all intron coordinates
}
