% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{sum_pip_annots}
\alias{sum_pip_annots}
\title{Sum PIPs for each annotation category}
\usage{
sum_pip_annots(finemap.locus.gr, annots.list, unassigned = "Intergenic")
}
\arguments{
\item{finemap.locus.gr}{fine mapping result for a locus}

\item{annots.list}{list of functional annotations}

\item{unassigned}{name of unassigned category, default: 'Intergenic'.}
}
\description{
Sum PIPs for each annotation category
}
