% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{susie_finemap_region}
\alias{susie_finemap_region}
\title{Run fine-mapping for one region with SuSiE using summary statistics (susie_rss)}
\usage{
susie_finemap_region(
  sumstats,
  R = NULL,
  snp_info = NULL,
  bigSNP = NULL,
  n = NULL,
  L = 1,
  prior_weights = NULL,
  estimate_residual_variance = FALSE,
  verbose = FALSE,
  save = FALSE,
  ...
)
}
\arguments{
\item{sumstats}{A data frame of summary statistics}

\item{R}{LD (correlation) matrix}

\item{snp_info}{Variant information for the LD matrix.}

\item{bigSNP}{a \code{bigsnpr} object attached via \code{bigsnpr::snp_attach()}
containing the reference genotype panel.}

\item{n}{GWAS sample size (optional, but strongly recommended.)}

\item{L}{Number of causal signals.}

\item{prior_weights}{A vector of prior probability for each SNP}

\item{estimate_residual_variance}{The default is FALSE,
the residual variance is fixed to 1 or variance of y.
If the in-sample LD matrix is provided,
we recommend setting estimate_residual_variance = TRUE.}

\item{verbose}{If TRUE, print progress}

\item{save}{If TRUE, save SuSiE result and LD (R) matrix.}
}
\value{
a list of SuSiE results, and z-scores and LD (R) matrix used in SuSiE
}
\description{
Run fine-mapping with SuSiE using summary statistics for one LD block
}
