% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{wrapped_distance}
\alias{wrapped_distance}
\title{Calculate distance between a point and references on a line}
\usage{
wrapped_distance(from, to, size)
}
\arguments{
\item{from}{the value we're measuring from}

\item{to}{one or more values to which we're measuring}

\item{size}{the length of the line}
}
\value{
a vector of distances to the values in "to"
}
\description{
Taking into account that the line wraps.
}
