% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{wrapped_range}
\alias{wrapped_range}
\title{Calculate wrapped ranges along a line.}
\usage{
wrapped_range(i, spread, size)
}
\arguments{
\item{i}{the centre of the range}

\item{spread}{the width of the range to each side of i}

\item{size}{the length of the line}
}
\value{
a vector of values within the range
}
\description{
Taking into account that the line wraps.
}
