% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{annotator}
\alias{annotator}
\title{Assigns SNPs with annotations based on overlap}
\usage{
annotator(gwas, annotations)
}
\arguments{
\item{gwas}{a data frame or tibble of GWAS summary statistics}

\item{annotations}{annotation BED files}
}
\description{
Assigns SNPs with annotations based on overlap
}
