% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{annotator_merged}
\alias{annotator_merged}
\title{Annotations for causal SNPs (apply these after fine-mapping!)}
\usage{
annotator_merged(gwas, annotations)
}
\arguments{
\item{gwas}{a data frame or tibble of GWAS summary statistics}

\item{annotations}{annotation BED files}
}
\description{
Annotations for causal SNPs (apply these after fine-mapping!)
}
