% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{assign_locus_snp}
\alias{assign_locus_snp}
\title{Assigns each SNP to one ld-block}
\usage{
assign_locus_snp(cleaned.sumstats, ld)
}
\arguments{
\item{cleaned.sumstats}{Cleaned summary statistics}

\item{ld}{LD blocks}
}
\description{
Assigns each SNP to one ld-block
}
