% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{assign_snp_locus}
\alias{assign_snp_locus}
\title{Assign GWAS SNPs to LD blocks}
\usage{
assign_snp_locus(sumstats, LD_Blocks)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics.}

\item{LD_Blocks}{A data frame of LD blocks with four columns,
'chr', 'start', 'end', and 'locus' (LD block indices).}
}
\value{
A data frame with summary statistics with assigned locus ID.
}
\description{
Assign GWAS SNPs to LD blocks
}
