% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{block_view_summary}
\alias{block_view_summary}
\title{LD block view summary table}
\usage{
block_view_summary(genemapping_res, finemapstats.gr)
}
\arguments{
\item{genemapping_res}{data frame of gene mapping result}

\item{finemapstats.gr}{GRange object of fine mapping result}
}
\description{
LD block view summary table
}
