% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{clean_sumstats}
\alias{clean_sumstats}
\title{Cleans summary statistics}
\usage{
clean_sumstats(
  sumstats,
  chr = "chr",
  pos = "pos",
  beta = "beta",
  se = "se",
  a0 = "a0",
  a1 = "a1",
  snp = "snp",
  pval = "pval"
)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics.}

\item{cols.to.keep}{columns to keep.
It is required to have 8 columns in the exact order of:
chr, position, beta, se, allele1, allele2, SNP ID (rs), p-value.}
}
\description{
Cleans summary statistics
}
