% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_sumstats}
\alias{clean_sumstats}
\title{Clean summary statistics}
\usage{
clean_sumstats(sumstats, cols.to.keep)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics.}

\item{cols.to.keep}{columns to keep in the data frame}
}
\description{
Clean summary statistics
}
