% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{compute_gene_pip}
\alias{compute_gene_pip}
\title{Compute gene PIPs based on fine-mapping result and functional annotations.}
\usage{
compute_gene_pip(
  finemapstats,
  genomic.annots,
  intron.mode = FALSE,
  d0 = 50000,
  cols.to.keep = c(names(mcols(finemapstats)), "gene_name", "category", "weight",
    "frac_pip", "gene_pip")
)
}
\arguments{
\item{finemapstats}{A GRanges object of fine-mapping result.}

\item{genomic.annots}{A list of GRanges objects of genomic annotations.}

\item{intron.mode}{Logical. If TRUE, assign intronic SNPs to genes containing the introns.}

\item{d0}{A scaling parameter used for computing weight based on SNP-gene distance.
Weight = exp(-dist/d0). Default = 50000 (50kb).}

\item{cols.to.keep}{columns to keep in the SNP gene weights}
}
\value{
A data frame of SNP-level view of gene mapping result
}
\description{
Compute gene PIPs based on fine-mapping result and functional annotations.
}
