% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{extract_gene_level_result}
\alias{extract_gene_level_result}
\title{Extract gene-level result from SNP-level gene mapping result}
\usage{
extract_gene_level_result(snp.gene.pip.mat, gene.annots)
}
\arguments{
\item{snp.gene.pip.mat}{A data frame of SNP-level gene mapping result}

\item{gene.annots}{a GRanges object of gene annotations}
}
\value{
a data frame of gene-level view of gene mapping result
}
\description{
Extract gene-level result from SNP-level gene mapping result
}
