% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{find_nearest_genes}
\alias{find_nearest_genes}
\title{Find the nearest genes for top SNPs in each locus.}
\usage{
find_nearest_genes(
  top.snps,
  genes,
  dist.to = c("genebody", "tss"),
  cols.to.keep = c("snp", "chr", "pos", "nearest_gene")
)
}
\arguments{
\item{top.snps}{A GRanges object of the GWAS summary statistics for the top SNPs}

\item{genes}{A GRanges object of gene information}

\item{dist.to}{Find nearest genes by distance to gene body or TSS}

\item{cols.to.keep}{columns to keep in the result}
}
\value{
A data frame with SNP locations and nearest genes.
}
\description{
Find the nearest genes for top SNPs in each locus.
}
