% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{find_nearest_genes}
\alias{find_nearest_genes}
\title{Find nearest genes (by distance to TSS or gene body)}
\usage{
find_nearest_genes(snp.gr, gene.gr, dist.to = c("tss", "genebody"))
}
\arguments{
\item{snp.gr}{GenomicRange object of SNP information}

\item{gene.gr}{GenomicRange object of gene information}

\item{dist.to}{Find nearest genes by distance to TSS or gene body}
}
\description{
Find nearest genes (by distance to TSS or gene body)
}
