% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{find_nearest_genes}
\alias{find_nearest_genes}
\title{Find nearest genes (by distance to TSS or gene body)}
\usage{
find_nearest_genes(
  gwas.gr,
  genes.gr,
  dist.to = c("genebody", "tss"),
  cols.to.keep = c("snp", "chr", "pos", "nearest_gene")
)
}
\arguments{
\item{gwas.gr}{GRanges object of the GWAS summary statistics}

\item{genes.gr}{GRanges object of gene information}

\item{dist.to}{Find nearest genes by distance to gene body or TSS.}

\item{cols.to.keep}{columns to keep in the returned GRanges object}
}
\description{
Find nearest genes (by distance to TSS or gene body)
}
