% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{gen_plates}
\alias{gen_plates}
\title{Generate tectonic plates}
\usage{
gen_plates(n_plates, map_x, map_y = map_x)
}
\arguments{
\item{n_plates}{number of plates on the map}

\item{map_x}{map x dimension in tiles}

\item{map_y}{map y dimension in tiles}
}
\value{
a list containing plate information, vector information, and the map
}
\description{
Generate tectonic plates
}
