% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{gene_by_distance}
\alias{gene_by_distance}
\title{Find all genes (promoters) within 1MB genes and assign weights based on distance}
\usage{
gene_by_distance(snps.gr, promoters.gr, c.dist = 50000)
}
\arguments{
\item{snps.gr}{a GRanges object with the SNP locations.}

\item{promoters.gr}{a GRanges object with the gene promoter locations.}

\item{c.dist}{A scaling number used for computing weight based on SNP-gene distance.
Weight = exp(-dist/c). Default = 50000 (50kb).}
}
\value{
a GRanges object with SNP to gene distance and weights calculated based on distance.
}
\description{
Find all genes (promoters) within 1MB genes and assign weights based on distance
}
