% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{gene_cs}
\alias{gene_cs}
\title{Get gene credible sets from SNP level gene mapping table}
\usage{
gene_cs(snp.gene.pip.mat, by.locus = TRUE, gene.cs.percent.thresh = 0.8)
}
\arguments{
\item{snp.gene.pip.mat}{A data frame of SNP level gene mapping table}

\item{by.locus}{Logical, if TRUE, get credible sets based on locus level gene PIP}

\item{gene.cs.percent.thresh}{percentage threshold for gene credible sets}
}
\description{
Get gene credible sets from SNP level gene mapping table
}
