% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gene_manhattan_plot}
\alias{gene_manhattan_plot}
\title{Make gene Manhattan plot}
\usage{
gene_manhattan_plot(
  gene.pip.res,
  chr = "chr",
  pos = "pos",
  gene_name = "gene_name",
  gene_pip = "gene_pip",
  gene.pip.thresh = 0.8,
  highlight = TRUE,
  ylim = 1.2,
  point.size = 2,
  label.size = point.size * 2,
  font.size = 15,
  max.overlaps = 20,
  title = ""
)
}
\arguments{
\item{gene.pip.res}{Gene mapping result}

\item{chr}{Name of the chr column in the gene mapping result}

\item{pos}{Name of the pos column in the gene mapping result}

\item{gene_name}{Name of the gene name column in the gene mapping result}

\item{gene_pip}{Name of the gene PIP column in the gene gene mapping result}

\item{gene.pip.thresh}{Gene PIP cutoff (default: 0.8)}

\item{highlight}{Highlight genes with gene PIP > gene.pip.thresh}

\item{ylim}{Truncate gene PIP to ylim value in the plot.}

\item{point.size}{Size of the points.}

\item{label.size}{Size of the labels.}

\item{font.size}{Font size of the text.}

\item{max.overlaps}{Exclude text labels that overlap too many things.}

\item{title}{Title of the plot}
}
\description{
Make gene Manhattan plot
}
