% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{gene_view_summary}
\alias{gene_view_summary}
\title{Gene view summary table}
\usage{
gene_view_summary(gene.mapping.res, gene.pip.thresh = 0.1)
}
\arguments{
\item{gene.mapping.res}{A data frame of gene mapping result}

\item{gene.pip.thresh}{Filter genes with gene PIP cutoff (default: 0.1)}
}
\description{
Gene view summary table
}
