% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_annots.R
\name{make_genomic_annots}
\alias{make_genomic_annots}
\title{Make genomic annotations from GTF file}
\usage{
make_genomic_annots(gtf.file, save = FALSE, outname = NULL)
}
\arguments{
\item{gtf.file}{GTF file name}

\item{save}{If TRUE, save genomic annotations as a RDS file}

\item{outname}{Output file name}
}
\description{
Make genomic annotations from GTF file
}
