% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.R
\name{make_genomic_annots}
\alias{make_genomic_annots}
\title{Make genomic annotations from a GTF file}
\usage{
make_genomic_annots(gtf_file, add_splice_junctions = FALSE)
}
\arguments{
\item{gtf_file}{Path to the GTF file}

\item{add_splice_junctions}{If TRUE, add splice junctions in the result}
}
\value{
A list of GRanges objects of genomic annotations
}
\description{
Make genomic annotations from a GTF file
}
