% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{merge_bigsnp_gwas}
\alias{merge_bigsnp_gwas}
\title{Match GWAS SNPs with reference panel from the bigSNP object.}
\usage{
merge_bigsnp_gwas(gwas, bigSNP)
}
\arguments{
\item{gwas}{GWAS summary statistics}

\item{bigSNP}{bigSNP object}
}
\description{
Match GWAS SNPs with reference panel from the bigSNP object.
}
