% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{merge_susie_sumstats}
\alias{merge_susie_sumstats}
\title{merges susie results with original summary statistics data frame}
\usage{
merge_susie_sumstats(susie_results, sumstats)
}
\arguments{
\item{susie_results}{data frame containing SuSiE finemapping result}

\item{sumstats}{data frame containing summary statistics}
}
\description{
merges susie results with original summary statistics data frame
Note. ASSUMES L = 1! ONLY ONE CREDIBLE SET PER LOCUS!
}
