% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{merge_susie_sumstats}
\alias{merge_susie_sumstats}
\title{merges SuSiE results with original summary statistics data frame}
\usage{
merge_susie_sumstats(susie_results, sumstats)
}
\arguments{
\item{susie_results}{SuSiE finemapping result}

\item{sumstats}{A data frame of summary statistics}
}
\value{
A data frame with summary statistics and two additional columns
of 'susie_pip' and 'cs' (CS indices)
}
\description{
merges SuSiE results with original summary statistics data frame
}
