% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.R
\name{nearby_interactions}
\alias{nearby_interactions}
\title{Get nearby interactions for enhancer regions near active promoters}
\usage{
nearby_interactions(
  enhancer_regions.gr,
  active_promoters.gr,
  dist_limit = 20000
)
}
\arguments{
\item{enhancer_regions.gr}{GRanges object of enhancer regions}

\item{active_promoters.gr}{GRanges object of active promoters}

\item{dist_limit}{Distance limit (default: 20kb)}
}
\description{
Get nearby interactions for enhancer regions near active promoters
}
