% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{partition_pip_annots}
\alias{partition_pip_annots}
\title{Partition PIPs into functional categories}
\usage{
partition_pip_annots(finemap.gr, annots, unassigned = "Intergenic")
}
\arguments{
\item{finemap.gr}{fine mapping result}

\item{annots}{functional annotations}

\item{unassigned}{name of unassigned category, default: 'Intergenic'.}
}
\description{
Partition PIPs into functional categories
}
