% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{partition_pip_regions}
\alias{partition_pip_regions}
\title{Partition PIPs into functional annotation regions}
\usage{
partition_pip_regions(finemap.gr, annots.list)
}
\arguments{
\item{finemap.gr}{a GRanges object of fine mapping summary statistics}

\item{annots.list}{a list of GRanges objects of functional annotations}
}
\value{
a list with a data frame with the sum of PIPs and
a data frame with the number of SNPs included for each annotation category.
}
\description{
Partition PIPs into functional annotation regions
}
