% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{partition_pip_regions}
\alias{partition_pip_regions}
\title{Partition PIPs into disjoint functional annotation regions}
\usage{
partition_pip_regions(finemapstats, annots.list)
}
\arguments{
\item{finemapstats}{A GRanges object of fine-mapping summary statistics}

\item{annots.list}{A list of GRanges objects of disjoint functional annotations.
Note: the annotation regions need to be disjoint!}
}
\value{
A list with a data frame with the sum of PIPs and
a data frame with the number of SNPs included for each annotation category.
}
\description{
Partition PIPs into disjoint functional annotation regions
}
