% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{prepare_susie_data_with_torus_result}
\alias{prepare_susie_data_with_torus_result}
\title{Prepare summary statistics with TORUS SNP-level priors as
input data for SuSiE}
\usage{
prepare_susie_data_with_torus_result(
  sumstats,
  torus_prior,
  torus_fdr,
  fdr.thresh = 0.1
)
}
\arguments{
\item{sumstats}{A data frame of summary statistics}

\item{torus_prior}{A data frame with SNP level priors
(result from \code{run_torus()} with \code{option=\dQuote{est-prior}})}

\item{torus_fdr}{A data frame containing the FDR of each region
(result from \code{run_torus()} with \code{option=\dQuote{fdr}}).
Optional, if available, only keep the loci with FDR < \code{fdr.thresh}.}

\item{fdr.thresh}{FDR cutoff (default: 0.1)}
}
\value{
A data frame of summary statistics with SNP-level priors
to be used as input data for \code{susie_rss}.
}
\description{
Check for required columns in summary statistics and
adds TORUS SNP-level priors to summary statistics to be used as input
data for \code{susie_rss}.
}
