% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{prepare_susie_data_with_torus_result}
\alias{prepare_susie_data_with_torus_result}
\title{Prepare SuSiE summary statistics with TORUS results}
\usage{
prepare_susie_data_with_torus_result(
  sumstats,
  torus_prior,
  torus_fdr,
  fdr_thresh = 0.1
)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics}

\item{torus_prior}{a tibble containing SNP level priors (result from run_torus)}

\item{torus_fdr}{a tibble containing the FDR of each region (result from run_torus)}
}
\value{
tibble of summary statistics updated with torus output
}
\description{
Adds TORUS results to the summary statistics
}
