% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{process_ABC}
\alias{process_ABC}
\title{Process ABC scores and save as a GRanges object}
\usage{
process_ABC(ABC.df, ABC.thresh = 0.015, full.element = TRUE, flank = 0)
}
\arguments{
\item{ABC.df}{a data frame of ABC scores}

\item{ABC.thresh}{Significance threshold of ABC scores (default = 0.015).}

\item{full.element}{Logical; if TRUE, use full length of ABC elements.
Otherwise, use the narrow regions provided in the ABC data.}

\item{flank}{Flanking regions around ABC elements (default = 0).}
}
\description{
Process ABC scores and save as a GRanges object
}
