% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_ABC}
\alias{process_ABC}
\title{Process ABC scores and save as a GRanges object}
\usage{
process_ABC(
  ABC,
  gene.annots = NULL,
  full.element = FALSE,
  score.thresh = 0.015,
  flank = 0
)
}
\arguments{
\item{ABC}{A data frame of ABC scores from Nasser et al. Nature 2021 paper}

\item{gene.annots}{If provided, restrict to genes in gene.annots}

\item{full.element}{Logical; if TRUE, use full length of ABC elements
extracted from the "name" column. Otherwise, use the original (narrow)
regions provided in the ABC scores data.}

\item{score.thresh}{Numeric. Threshold of ABC scores.
(default = 0.015, as in Nasser et al. Nature 2021 paper).}

\item{flank}{Integer. Extend bases on both sides of the ABC elements (default = 0).}
}
\value{
a GRanges object with processed ABC scores, with genomic coordinates
of the interacting regions and gene names (promoters).
}
\description{
Process ABC scores and save as a GRanges object
}
