% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{process_gwas_sumstats}
\alias{process_gwas_sumstats}
\title{Cleans GWAS summary statistics and adds metadata}
\usage{
process_gwas_sumstats(
  sumstats_file,
  chr = "chr",
  pos = "pos",
  beta = "beta",
  se = "se",
  a0 = "a0",
  a1 = "a1",
  snp = "snp",
  pval = "pval",
  bigSNP,
  LD_Blocks
)
}
\arguments{
\item{sumstats_file}{file containing raw summary statistics.
Coordinates should be hg19/b37; we current do not support hg38.}

\item{chr}{Name of the chromosome column in summary statistics file.}

\item{pos}{Name of the position column (base pair position, in \emph{hg19}!).}

\item{beta}{Name of beta column (if you have Odds Ratio,
you will need to transform it to log(Odds Ratio)).}

\item{se}{Name of the standard error (se) column.}

\item{a0}{Column name of the reference allele.}

\item{a1}{Column name of the association/effect allele.}

\item{snp}{Name of the SNP ID (rsID) column.}

\item{pval}{Name of the p-value column.}

\item{bigSNP}{bigSNP object from \code{bigsnpr}.}

\item{LD_Blocks}{Reference LD blocks}
}
\value{
Cleaned summary statistics + LD block of every SNP,
as well as its index in the reference panel of genotypes
}
\description{
Cleans GWAS summary statistics and adds metadata
}
