% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{process_gwas_sumstats}
\alias{process_gwas_sumstats}
\title{Cleans GWAS summary statistics and adds metadata}
\usage{
process_gwas_sumstats(
  sumstats,
  chr = "chr",
  pos = "pos",
  beta = "beta",
  se = "se",
  a0 = "a0",
  a1 = "a1",
  snp = "snp",
  pval = "pval",
  remove_indels = TRUE,
  LD_Blocks,
  bigSNP,
  ...
)
}
\arguments{
\item{sumstats}{A data frame of GWAS summary statistics.}

\item{chr}{Name of the chromosome column in summary statistics.}

\item{pos}{Name of the position column (base pair position).}

\item{beta}{Name of beta column (if you have Odds Ratio,
you will need to transform it to log(Odds Ratio)).}

\item{se}{Name of the standard error (se) column.}

\item{a0}{Column name of the reference allele.}

\item{a1}{Column name of the association/effect allele.}

\item{snp}{Name of the SNP ID (rsID) column.}

\item{pval}{Name of the p-value column.}

\item{remove_indels}{If TRUE, remove indels}

\item{LD_Blocks}{A data frame of LD blocks with four columns,
'chr', 'start', 'end', and 'locus' (LD block indices).}

\item{bigSNP}{a \code{bigsnpr} object attached via \code{bigsnpr::snp_attach()}
containing the reference genotype panel.}
}
\value{
A data frame of processed GWAS summary statistics.
}
\description{
Cleans GWAS summary statistics and adds metadata, including:
the index in the bigSNP object of each SNP, and the LD block locus.
}
