% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pcHiC}
\alias{process_pcHiC}
\title{Process PC-HiC data and save as a GRanges object}
\usage{
process_pcHiC(pcHiC, gene.annots, score.thresh = 0, flank = 0)
}
\arguments{
\item{pcHiC}{A data frame of PC-HiC data,
with columns of enhancer"Promoter" and
"Interacting_fragment". Interacting_fragment should contains
chr, start and end positions of the fragments interacting with promoters
e.g. "chr.start.end" or "chr:start-end".}

\item{gene.annots}{If provided, restrict to genes in gene.annots}

\item{score.thresh}{Numeric. Threshold of interaction scores.
(default = 0).}

\item{flank}{Integer. Extend bases on both sides of the regulatory elements
(default = 0).}
}
\value{
A GRanges object with processed PC-HiC links, with genomic coordinates
of the interacting regions and gene names (promoters).
}
\description{
Process PC-HiC data and save as a GRanges object
}
