% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{process_pcHiC}
\alias{process_pcHiC}
\title{Process pcHiC data and save as a GRanges object}
\usage{
process_pcHiC(pcHiC)
}
\arguments{
\item{pcHiC}{a data frame of pcHiC data}
}
\description{
Process pcHiC data and save as a GRanges object
}
