% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_finemapping}
\alias{run_finemapping}
\title{Run fine-mapping using GWAS summary statistics}
\usage{
run_finemapping(
  sumstats,
  bigSNP,
  n,
  priortype = c("torus", "uniform"),
  L = 1,
  estimate_residual_variance = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sumstats}{A data frame containing raw summary statistics; must have header!}

\item{bigSNP}{a bigsnpr object attached via bigsnpr::snp_attach()}

\item{n}{The sample size}

\item{priortype}{prior type: "torus" or "uniform".}

\item{L}{Number of causal signals}

\item{estimate_residual_variance}{The default is FALSE,
the residual variance is fixed to 1 or variance of y.
If the in-sample LD matrix is provided,
we recommend setting estimate_residual_variance = TRUE.}

\item{verbose}{If verbose = TRUE, print progress,
and a summary of the optimization settings from susie.}
}
\value{
list of finemapping results; one per LD block
}
\description{
Run finemapping with SuSiE using GWAS summary statistics
for all LD blocks with prior probabilities computed by TORUS
}
