% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_finemapping}
\alias{run_finemapping}
\title{Run finemapping}
\usage{
run_finemapping(sumstats, bigSNP, priortype = c("torus", "uniform"), L = 1)
}
\arguments{
\item{sumstats}{a tibble or data frame containing raw summary statistics; must have header!}

\item{bigSNP}{a bigsnpr object attached via bigsnpr::snp_attach()}

\item{priortype}{prior type: "torus" or "uniform".}

\item{L}{Number of causal signals}
}
\value{
list of finemapping results; one per LD block
}
\description{
Run finemapping with SuSiE for all LD blocks
}
