% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_susie}
\alias{run_susie}
\title{run SUSIE}
\usage{
run_susie(sumstats, bigSNP, ldchunk, L, prior)
}
\arguments{
\item{sumstats}{summary statistics}

\item{bigSNP}{bigSNP object}

\item{ldchunk}{LD chunk}

\item{L}{Number of causal signals}

\item{prior}{Logical, if TRUE, use the \code{torus_pip} column
in \code{sumstats} as prior}
}
\description{
run SUSIE
}
