% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{run_susie}
\alias{run_susie}
\title{Run fine-mapping with SUSIE for one LD block}
\usage{
run_susie(sumstats, bigSNP, locus, L = 1, useprior)
}
\arguments{
\item{sumstats}{summary statistics}

\item{bigSNP}{bigSNP object}

\item{locus}{LD block index}

\item{L}{Number of causal signals}

\item{useprior}{Logical, if TRUE, use the \code{torus_prior} column
in \code{sumstats} as prior.}
}
\value{
finemapping results for one LD block
}
\description{
Run fine-mapping with SUSIE for one LD block
}
