% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{snp_view_summary}
\alias{snp_view_summary}
\title{SNP view summary table}
\usage{
snp_view_summary(
  gene.mapping.res,
  gene.annots,
  finemapstats,
  fractional.pip.thresh = 0.02
)
}
\arguments{
\item{gene.mapping.res}{A data frame of gene mapping result}

\item{gene.annots}{A data frame of gene annotations}

\item{finemapstats}{A GRange object of fine-mapping summary statistics}

\item{fractional.pip.thresh}{Filter SNPs with fractional PIP cutoff (default: 0.02)}
}
\value{
A data frame of SNP view summary of gene mapping result
}
\description{
SNP view summary table
}
