% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{structure_plot}
\alias{structure_plot}
\title{Make a structure plot of partitioned PIP by locus}
\usage{
structure_plot(
  mat,
  categories,
  colors,
  ticks = NULL,
  font.size = 9,
  highlight = NULL,
  xlab = "Locus",
  ylab = "Proportion",
  legend.title = "Category"
)
}
\arguments{
\item{mat}{matrix of input data, rows are loci, columns are annotation categories}

\item{categories}{annotation categories}

\item{colors}{Colors of the structure plot categories}

\item{ticks}{Labels of x-axis}

\item{font.size}{Font size of structure plot}

\item{highlight}{Highlight a locus}
}
\description{
Making a structure plot of partitioned PIP by locus
This function is adapted from the 'fastTopics' package
https://stephenslab.github.io/fastTopics/
}
