% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_pips.R
\name{sum_pip_annots}
\alias{sum_pip_annots}
\title{Sum PIPs for each annotation category,
assign SNPs to annotations with the priority of the earlier category}
\usage{
sum_pip_annots(finemapstats.locus.gr, annots.list)
}
\description{
Sum PIPs for each annotation category,
assign SNPs to annotations with the priority of the earlier category
}
