% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{track_plot}
\alias{track_plot}
\title{Make gene track plot using Gviz}
\usage{
track_plot(
  finemapstats,
  region,
  gene.annots,
  bigSNP,
  txdb,
  counts,
  peaks,
  loops,
  genome = "hg19",
  filter_loop_genes = NULL,
  filter_loop_snps = NULL,
  filter_protein_coding_genes = TRUE,
  color_piptrack_by = c("locus", "cs", "none"),
  highlight_snps = NULL,
  counts.ylim = c(0, 1),
  counts.color,
  peaks.color = "navy",
  loops.color = "gray",
  highlight.color = "pink",
  genelabel.side = c("above", "below", "left", "right"),
  track.sizes,
  rotation.title = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{finemapstats}{A data frame of finemapping summary statistics}

\item{region}{A GRanges object or data frame for the genomic range to plot}

\item{gene.annots}{A GRanges object of gene annotations}

\item{bigSNP}{A \code{bigsnpr} object attached via bigsnpr::snp_attach()}

\item{txdb}{A \code{txdb} object of gene annotations}

\item{counts}{A list of counts data to display as quantitativ data tracks}

\item{peaks}{A list of peaks to display as binary data tracks}

\item{loops}{A list of chromatin loops, e.g. PC-HiC, ABC, etc.}

\item{genome}{Genome assembly version (default: "hg19").}

\item{filter_loop_genes}{A vector of gene names. Only show loops connected to the genes.}

\item{filter_loop_snps}{A vector of SNP IDs. Only show loops connected to the SNPs.}

\item{filter_protein_coding_genes}{Logical. If TRUE, only shows protein coding gene.}

\item{color_piptrack_by}{Color SNPs in the PIP track by
\code{locus}, \code{cs} (credible sets), or \code{none} (same color).}

\item{highlight_snps}{SNPs (rsIDs) to highlight. Or highlight top SNP ('topSNP').}

\item{counts.ylim}{ylim range (default: between 0 and 1) for the \code{counts} tracks}

\item{counts.color}{Colors for the \code{counts} tracks}

\item{peaks.color}{Colors for the \code{peak} tracks}

\item{highlight.color}{Colors for the highlighted SNPs}

\item{genelabel.side}{Side to put gene labels,
options are: 'above' (default), 'below', 'left', 'right'.}

\item{track.sizes}{Sizes of the tracks}

\item{rotation.title}{The rotation angle for the text in the track title}

\item{verbose}{If TRUE, print detail messages for plotting}
}
\description{
Make gene track plot using Gviz
}
